feat: 유저 상태에 따라 chatroom 컴포넌트 분리 (#103)
